Set up derivatives tests
